refactor: Introduce Experiment mixin
